requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
colorama==0.4.6
coverage==7.2.3
Django==4.1.7
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.1
Faker==18.3.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mypy==1.1.1
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
parameterized==0.9.0
Pillow==9.4.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
pytest==7.3.2
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==2.0.2
watchdog==3.0.0
wcwidth==0.2.6
wsproto==1.2.0
